refactor: refactor tables to use named export
