--- src/components/ui/Tables/DestinationsTable/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./DestinationsTable";
-export * from "./DestinationsTable";
