--- src/components/ui/Tables/ModelsTable/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ModelsTable";
-export * from "./ModelsTable";
