--- src/components/ui/Tables/PipelineTable/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./PipelineTable";
-export * from "./PipelineTable";
