--- src/components/ui/Tables/PipelinesTable/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./PipelinesTable";
-export * from "./PipelinesTable";
