--- src/components/ui/Tables/SourcesTable/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./SourcesTable";
-export * from "./SourcesTable";
